fixed dismissing issues
spikes where not dismissing it was fixed.
next update I will be to where tires don't explode 
as of now it is fivem key bindable. preset keys are in the readme
been just getting back into coding so if you like and download thanks for the love. and please give feedback!